requirements.txt
AfricastalkingGateway==1.8
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.4
django-leaflet==0.22.0
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
urllib3==1.22
whitenoise==3.3.0
